fix time_idx dtype

+++ link_bot_planning/src/link_bot_planning/results_to_dynamics_dataset.py
@@ -163,7 +163,7 @@ class ResultsToDynamicsDataset:
             actions_dict = sequence_of_dicts_to_dict_of_np_arrays(action_subsequence, 0)
             states_dict = sequence_of_dicts_to_dict_of_np_arrays(state_subsequence, 0)
 
-            time_idx = np.arange(len(state_subsequence))
+            time_idx = np.arange(len(state_subsequence), dtype=np.float32)
             environment = steps[0]['planning_query'].environment
             example = {
                 'traj_idx': trial_idx,
